image/remove: emulate docker's untagging
When the Docker CLI removes an image:tag it effectively removes the tag.
This change emulates such behavior but will fallback to the CRI
implementation if an "image not found" error is returned from the
containerd-native delete. This is typically only the case when passing
image identifiers (truncated or otherwise) incurs behavior equivalent to
`docker image rm --force --no-prune <image-id>`.

Signed-off-by: Jacob Blain Christen <jacob@rancher.com>