fix agent crash when removing invalid image (#83)
The client attempts to "noramlize" image refs when deleting to allow for
shortened image identifiers to be passed. When passing an invalid image
ref this could result in the client requesting deletion of a nil image
from the server. The server now silently succeed and do nothing while
the client now checks for passing a nil image ref and lets the user know
that no such image could be found.

Fixes #80

Signed-off-by: Jacob Blain Christen <jacob@rancher.com>